Ensure that messages from the same party are assigned to the same topic partition.
This was already the intention; however we pass the key in as a byte array, and Kafka uses Object.hashCode() to determine the partition. This goes wrong, because byte arrays don't have a semantic hashcode: the identity is used.
The fix is to pass in the string explicitly as the partition key.

(cherry picked from commit 53a84d4d354a7ae0f4486d9e275f23c148b3567b)