Don't cache elements during script initialisation: the DOM may not be ready yet.
Elements are used in some browsers to determine the window dimensions. This commit switches to fetching the underlying elements (if necessary) during each signal, instead of caching them during script initialisation. (There might still be a problem is signals are issued prior to the DOM being ready. This is something to look into.)