Write absolute mapping script path to log.
This changes the diagnostic logging when locating a mapping script to catch a
subtle misconfiguration that's difficult to spot without this patch.

Specifically, this configuration:

    divolte.mappings.x {
        mapping_script_file = 'some/script.groovy'
    }

Is materialised into this filesystem path:

    /path/to/divolte/'some/script.groovy'

The configuration format supports double-quoted strings but not single-quoted
strings. By writing the java.io.File's absolute path to the log file instead of
the original configuration value, this becomes easier to spot.