Fix a potential deadlock while receiving asynchronous request bodies.
The deadlock could occur when the maximum buffer size (either configured, or indicated by the content-length header) was an exact multiple of the internal chunk size (4096 bytes). In this case, the last read may or may not detect EOF:
 - If it hasn’t, everything worked fine: reads resumed and the eventual EOF triggered the async callback, leading to normal completion.
 - If it has, then resuming reads won’t trigger any more callbacks and the lifecycle never completes.

The fix is to always try to read EOF immediately, and only resume async reading if EOF has not yet been detected.

For some reason one of the unit tests started triggering this bug in Undertow 1.4.7. It’s possible that this release changed the semantics of how resumable channels work, but the release notes don’t mention anything and I suspect the possibility was always there.