Change port number selection for the test server.
Instead of using random ports, we now cycle through a white-list of port numbers. This is because browsers block most ports on localhost for security reasons, which causes our browser-based tests to fail unless they happen to land on a port that’s on the whitelist.

In addition, Sauce Labs only forward a specific set of ports from localhost through to the Sauce Connect proxy. Connections to anything not on this list fail.