Update test server so that stopping is done in the background.
Stopping the server takes some time: shutting down the internal queues always waits until the next heartbeat occurs. Given that normally we’re stopping the server during test teardown, we don’t need to wait for it to complete. Instead we can trigger the shutdown and let it proceed in the background.