Use the Page Transition API to detect when page-view events should be fired.
This means we now work properly when bfcache is active if browsers support it. Some changes this necessitates include:
 - Deferring generation of the pageViewId until when the page-view starts.
 - Detecting when a new page-view should commence after the initial load.

A workaround is also present for old Opera browsers which use bfcache but don’t support the Page Transition API. In this case we disable Opera’s bfcache as a last resort.