Fix free port detection in the test server.
The code which detects free ports didn't work properly because it didn't
take into account the host on which we're binding (localhost or
external). Previously it probed against IN_ADDR_ANY, which is a different
port-space to the interface-specific ports that we're actually using.
The impact was spurious BindExceptions being thrown during tests.